relax JRE requirement for RXTX library from Java 7 to Java 5
After this change the jbidibc-rxtx jar is usable in Java 5 and
Java 6 JREs and JDKs.  Without it Java 7 is required.

Both 5 and 6 are already end-of-life, but there are no Java 6 or
Java 7 features used in the RXTX package and it might as well be
usable for users stuck with an old JVM.